refactor: typed parser state machine with per-block builders (TB-6)
`parse_page` block-in-progress state lives in a typed `ParseState` enum
(parser/state.rs): `Idle | Paragraph(ParagraphBuilder) | List(ListBuilder) |
CodeBlock(CodeBlockBuilder)`. Each variant owns its in-progress builder;
`flush_in_place` rotates the state through `Idle` and emits the completed
block plus any structural diagnostic that goes with it (e.g.
`parse.unclosed_fence` from a code block that hits EOF without a closing
fence).

The previous tuple of mutable variables (`paragraph_lines`,
`paragraph_start_line`, `paragraph_end_line`, `pending_list`) and the inline
code-fence sub-loop collapse into one state object whose transitions are
explicit. The orchestrator becomes a small per-line classifier that hands
off to focused consumers (`consume_heading`, `consume_list_item`,
`consume_prose_line`, `consume_code_block_line`) sharing a `LineContext`
bundle so no helper exceeds clippy's argument cap.

Module split per ADR-0007/ADR-0006 maintenance pressure: `parser/{mod,
state, builders}.rs`. mod.rs holds orchestration + heading/annotation/
ordered-list-prefix recognition; state.rs holds the enum and `flush`;
builders.rs holds the three block-builders (ParagraphBuilder, ListBuilder
moved from TB-5, CodeBlockBuilder new).

ADR-0004 gains an addendum documenting the state-machine decision and why
it preserves the "small parse functions" intent. TB-1 snapshots remain
green; the list-span pin in tests still passes (ListBuilder deliberately
does not extend list_span on push — TB-9 will switch the validator to
walk per-item spans instead).
